add vf.load_environment to orchestrator, add pydantic typing for vf, move registry to environments folder
