+++ configs/endpoints.py
@@ -1,0 +1,42 @@
+ENDPOINTS = {
+    "my-model": {
+        "model": "my-model",
+        "url": "https://some-endpoint.com/v1",
+        "key": "SOME_API_KEY",
+    },
+    "gpt-4.1-mini": {
+        "model": "gpt-4.1-mini",
+        "url": "https://api.openai.com/v1",
+        "key": "OPENAI_API_KEY",
+    },
+    "gpt-4.1-nano": {
+        "model": "gpt-4.1-nano",
+        "url": "https://api.openai.com/v1",
+        "key": "OPENAI_API_KEY",
+    },
+    "gpt-4.1": {
+        "model": "gpt-4.1",
+        "url": "https://api.openai.com/v1",
+        "key": "OPENAI_API_KEY",
+    },
+    "deepseek-v3": {
+        "model": "deepseek-chat",
+        "url": "https://api.deepseek.com/v1",
+        "key": "DEEPSEEK_API_KEY",
+    },
+    "qwen3-instruct": {
+        "model": "Qwen/Qwen3-235B-A22B-Instruct-2507",
+        "url": "https://api.deepinfra.com/v1/openai",
+        "key": "DEEPINFRA_API_KEY",
+    },
+    "qwen3-coder": {
+        "model": "Qwen/Qwen3-Coder-480B-A35B-Instruct",
+        "url": "https://api.deepinfra.com/v1/openai",
+        "key": "DEEPINFRA_API_KEY",
+    },
+    "kimi-k2": {
+        "model": "moonshotai/Kimi-K2-Instruct",
+        "url": "https://api.deepinfra.com/v1/openai",
+        "key": "DEEPINFRA_API_KEY",
+    },
+}
