--- environments/taubench2/README.md
@@ -1,1 +1,0 @@
-# taubench2
