+++ environments/vf_ascii_tree/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "vf-ascii-tree"
+version = "0.1.0"
+dependencies = [
+    "verifiers>=0.1.2",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build]
+include = ["vf_ascii_tree.py"]
