+++ environments/vf_hendrycks_math/pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "vf-hendrycks-math"
+version = "0.1.0"
+dependencies = [
+    "verifiers>=0.1.2",
+    "datasets",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build]
+include = ["vf_hendrycks_math.py"]
