+++ environments/vf_intellect_math/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "vf-intellect-math"
+version = "0.1.0"
+dependencies = [
+    "verifiers>=0.1.2",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build]
+include = ["vf_intellect_math.py"]
