+++ environments/vf_pydantic_adherence/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "vf-pydantic-adherence"
+version = "0.1.0"
+dependencies = [
+    "verifiers>=0.1.2",
+    "datasets",
+    "pydantic",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build]
+include = ["vf_pydantic_adherence.py"]
