+++ environments/vf_skywork_math/vf_skywork_math.py
@@ -1,0 +1,41 @@
+import json
+
+import verifiers as vf
+from datasets import load_dataset
+
+from prime_rl.orchestrator.genesys.math import compute_math_reward
+
+
+def load_environment(
+    solve_rate_field: str | None = None,
+    min_solve_rate: float | None = None,
+    max_solve_rate: float | None = None,
+) -> vf.Environment:
+    train_dataset = load_dataset("PrimeIntellect/Skywork-OR1-RL-Data-v1-math-prime-rl-format", split="train").map(
+        lambda x: {
+            "question": x["prompt"],
+            "info": json.loads(x["verification_info"]),
+            "task": "simple-math",
+        }
+    )
+    if solve_rate_field is not None:
+        if min_solve_rate is not None:
+            train_dataset = train_dataset.filter(lambda x: x[solve_rate_field] >= min_solve_rate)
+        if max_solve_rate is not None:
+            train_dataset = train_dataset.filter(lambda x: x[solve_rate_field] <= max_solve_rate)
+    train_dataset = train_dataset.remove_columns(["prompt", "verification_info"])
+    train_dataset = train_dataset.shuffle(seed=42)
+
+    def correct_answer_reward_func(completion, info) -> float:
+        completion_text = completion[-1]["content"]
+        return compute_math_reward(completion_text, info)
+
+    rubric = vf.Rubric(
+        funcs=[
+            correct_answer_reward_func,
+        ],
+        weights=[1.0],
+    )
+
+    vf_env = vf.SingleTurnEnv(dataset=train_dataset, rubric=rubric)
+    return vf_env
