+++ environments/vf_unscramble/pyproject.toml
@@ -1,8 +1,9 @@
 [project]
-name = "xlam-function-calling"
+name = "vf-unscramble"
 version = "0.1.0"
 dependencies = [
     "verifiers>=0.1.2",
+    "datasets",
 ]
 
 [build-system]
@@ -10,4 +11,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build]
-include = ["xlam_function_calling.py"]
+include = ["vf_unscramble.py"]
