--- environments/xlam_function_calling/README.md
@@ -1,1 +1,0 @@
-# xlam-function-calling
