--- environments/xlam_function_calling/xlam_function_calling.py
@@ -1,8 +1,0 @@
-import verifiers as vf
-
-
-def load_environment(**kwargs) -> vf.Environment:
-    """
-    Loads a custom environment.
-    """
-    raise NotImplementedError("Implement your custom environment here.")
