+++ pyproject.toml
@@ -31,6 +31,8 @@ dependencies = [
     "verifiers",
     "textarena>=0.6.16",
     "nltk>=3.9.1",
+    "vf-intellect-math",
+    "vf-wordle",
 ]
 
 [project.scripts]
@@ -54,12 +56,19 @@ prerelease = "allow"
 
 [tool.uv.sources]
 torch = [{ index = "pytorch-cu128" }]
-verifiers = { git = "https://github.com/willccbb/verifiers", rev = "08016c4" }
+verifiers = { git = "https://github.com/willccbb/verifiers", rev = "75c7df8" }
+vf-intellect-math = { workspace = true }
+vf-wordle = { git = "https://github.com/willccbb/verifiers.git", subdirectory = "environments/vf_wordle", rev = "main" }
 
 [[tool.uv.index]]
 name = "pytorch-cu128"
 url = "https://download.pytorch.org/whl/test/cu128"
 explicit = true
+
+[tool.uv.workspace]
+members = [
+    "environments/vf_intellect_math",
+]
 
 [dependency-groups]
 dev = [
