+++ src/prime_rl/orchestrator/orchestrator.py
@@ -10,11 +10,11 @@ from prime_rl.orchestrator import envs
 import lovely_tensors as lt
 import numpy as np
 import torch
+from verifiers import load_environment
 from transformers import AutoTokenizer
 
 from prime_rl.eval.utils import run_benchmark
 from prime_rl.orchestrator.ckpt import CheckpointManager, Progress
-from prime_rl.environments.registry import load_environment
 from prime_rl.orchestrator.client import (
     check_has_model,
     check_health,
@@ -77,7 +77,7 @@ async def orchestrate(config: OrchestratorConfig):
     ckpt_step = 0
     if config.ckpt and config.ckpt.resume_step:
         logger.info(f"Resuming training from checkpoint step `{config.ckpt.resume_step}`")
-        ckpt_manager.load(progress, step=config.ckpt.resume_step)
+        ckpt_manager.load(progress, step=config.ckpt.resume_step)  # type: ignore
         ckpt_step = max(progress.step - config.async_level, 0)
         await reload_weights(client, config.weights_path, ckpt_step)
     else:
@@ -86,7 +86,7 @@ async def orchestrate(config: OrchestratorConfig):
 
     # Load environment and extract dataset
     logger.info(f"Loading environment {config.environment.id} with args {config.environment.args}")
-    vf_env = load_environment(config.environment.id, config.environment.args)
+    vf_env = load_environment(config.environment.id, **config.environment.args)
     dataset = vf_env.get_dataset(seed=config.seed)
 
     # Setup buffer
@@ -108,7 +108,7 @@ async def orchestrate(config: OrchestratorConfig):
         if config.ckpt and config.ckpt.interval and not is_first_step and progress.step % config.ckpt.interval == 0:
             logger.info(f"Saving checkpoint at step {progress.step}")
             save_ckpt_start_time = time.time()
-            ckpt_manager.save(progress, step=progress.step)
+            ckpt_manager.save(progress, step=progress.step)  # type: ignore
             save_ckpt_time = time.time() - save_ckpt_start_time
 
         # Break if we have reached the maximum number of steps
@@ -210,10 +210,10 @@ async def orchestrate(config: OrchestratorConfig):
             generate_completions_time = time.time() - generate_completions_start_time
 
             results = vf_env.process_env_results_vllm(
-                prompts=outputs["prompt"],
-                completions=outputs["completion"],
-                states=outputs["state"],
-                rewards=outputs["reward"],
+                prompts=outputs.prompt,
+                completions=outputs.completion,
+                states=outputs.state,
+                rewards=outputs.reward,
                 processing_class=tokenizer,
                 max_seq_len=config.seq_len,
                 mask_env_responses=config.mask_env_responses,
@@ -222,7 +222,7 @@ async def orchestrate(config: OrchestratorConfig):
             )
 
             advantages = compute_advantages(
-                rewards=outputs["reward"],
+                rewards=outputs.reward,
                 samples_per_problem=config.rollouts_per_prompt,
                 advantage_type=config.advantage_type,
             )
@@ -230,12 +230,12 @@ async def orchestrate(config: OrchestratorConfig):
             # Update pool
             rollouts = make_rollouts(
                 problem_ids=problem_ids,
-                prompt_tokens=results["prompt_ids"],
-                prompt_masks=results["prompt_mask"],
-                completion_tokens=results["completion_ids"],
-                completion_masks=results["completion_mask"],
-                completion_logprobs=results["completion_logprobs"],
-                rewards=outputs["reward"],
+                prompt_tokens=results.prompt_ids,
+                prompt_masks=results.prompt_mask,
+                completion_tokens=results.completion_ids,
+                completion_masks=results.completion_mask,
+                completion_logprobs=results.completion_logprobs,
+                rewards=outputs.reward,
                 advantages=advantages,
             )
             buffer.update(rollouts)
@@ -381,7 +381,7 @@ async def orchestrate(config: OrchestratorConfig):
     # Write final checkpoint
     if config.ckpt:
         logger.info("Writing final checkpoint")
-        ckpt_manager.save(progress, step=progress.step)
+        ckpt_manager.save(progress, step=progress.step)  # type: ignore
 
     logger.success("Orchestrator finished.")
